moved the component to diff file
