+++ src/containers/Template/Form/HSM/HSM.tsx
@@ -12,6 +12,7 @@ import { Loading } from 'components/UI/Layout/Loading/Loading';
 import { Simulator } from 'components/simulator/Simulator';
 import Template from '../Template';
 import styles from './HSM.module.css';
+import { TemplateVariables } from '../../TemplateVariables/TemplateVariables';
 
 const defaultAttribute = {
   isHsm: true,
@@ -138,6 +139,9 @@ export const HSM = () => {
       onChange: (event: any) => {
         setCategory(event);
       },
+    },
+    {
+      component: TemplateVariables,
     },
   ];
