+++ src/containers/Template/Form/Template.tsx
@@ -27,7 +27,7 @@ import {
 import { CreateAutoComplete } from 'components/UI/Form/CreateAutoComplete/CreateAutoComplete';
 import { validateMedia } from 'common/utils';
 import styles from './Template.module.css';
-import { TemplateVariables } from './TemplateVariables';
+import { TemplateVariables } from '../TemplateVariables/TemplateVariables';
 
 const regexForShortcode = /^[a-z0-9_]+$/g;
 
@@ -689,9 +689,6 @@ const Template = ({
       },
       isEditing: isEditing,
       editorState: editorState,
-    },
-    {
-      component: TemplateVariables,
     },
   ];
