+++ src/containers/Template/TemplateVariables/TemplateVariable.module.css
@@ -1,0 +1,10 @@
+.AddVariable {
+  border-radius: 8px;
+}
+
+.AddIcon {
+  height: 80%;
+  text-transform: capitalize;
+  border-radius: 20px;
+  color: #119656;
+}
